Warning...
git-svn-id: svn://svn.code.sf.net/p/tass64/code/trunk@2634 111d7451-93fa-4e81-90d4-8039f76616e9

+++ main.c
@@ -66,7 +66,7 @@ static inline void install_signal_handler(void) {
 #if defined _POSIX_C_SOURCE || _POSIX_VERSION >= 199506L
     struct sigaction sa, osa;
     sa.sa_handler = signal_handler;
-    sa.sa_flags = SA_RESETHAND;
+    sa.sa_flags = (int)SA_RESETHAND;
     sigemptyset(&sa.sa_mask);
     if (sigaction(SIGINT, NULL, &osa) == 0 && osa.sa_handler != SIG_IGN) {
         sigaction(SIGINT, &sa, NULL);
